Refactor TorchDeprecatedSymbolsVisitor
